+++ src/wallet/test/wallet_tests.cpp
@@ -598,7 +598,7 @@ public:
         std::vector<COutPoint> vecOutpoints;
         size_t n;
         for (n = 0; n < tx->vout.size(); ++n) {
-            if (nChangePosRet != -1 && n == nChangePosRet) {
+            if (nChangePosRet != -1 && int(n) == nChangePosRet) {
                 // Skip the change output to only return the requested coins
                 continue;
             }
@@ -632,7 +632,7 @@ BOOST_FIXTURE_TEST_CASE(CreateTransactionTest, CreateTransactionTestSetup)
             {13, {{100001, true}}}
         };
         assert(mapTestCases.size() == mapExpected.size());
-        for (int i = 0; i < mapTestCases.size(); ++i) {
+        for (size_t i = 0; i < mapTestCases.size(); ++i) {
             if (!CreateTransaction(mapTestCases.at(i), mapExpected.at(i).first, mapExpected.at(i).second)) {
                 std::cout << strprintf("CreateTransactionTest failed at: %d - %d\n", nTestId, i) << std::endl;
             }
@@ -858,10 +858,10 @@ BOOST_FIXTURE_TEST_CASE(CreateTransactionTest, CreateTransactionTestSetup)
         // Just to create nCount output recipes to use in tests below
         std::vector<std::pair<CAmount, bool>> vecOutputEntries{{5000, false}};
         auto createOutputEntries = [&](int nCount) {
-            while (vecOutputEntries.size() <= nCount) {
+            while (vecOutputEntries.size() <= size_t(nCount)) {
                 vecOutputEntries.push_back(vecOutputEntries.back());
             }
-            if (vecOutputEntries.size() > nCount) {
+            if (vecOutputEntries.size() > size_t(nCount)) {
                 int nDiff = vecOutputEntries.size() - nCount;
                 vecOutputEntries.erase(vecOutputEntries.begin(), vecOutputEntries.begin() + nDiff);
             }
